remove unfoldNd dependence
