+++ involution/__init__.py
@@ -1,2 +1,3 @@
 # Import involution module
 from .involution import Involution2d
+from .involution import Involution3d
