--- unfoldNd/__init__.py
@@ -1,12 +1,0 @@
-"""unfoldNd library."""
-
-
-from unfoldNd.fold import FoldNd, foldNd
-from unfoldNd.unfold import UnfoldNd, unfoldNd
-
-__all__ = [
-    "UnfoldNd",
-    "unfoldNd",
-    "FoldNd",
-    "foldNd",
-]
